int relaxation time

+++ src/qibolab/platforms/platform.py
@@ -72,7 +72,7 @@ class ExecutionParameters:
     """
 
     nshots: Optional[int] = 1024
-    relaxation_time: Optional[float] = None
+    relaxation_time: Optional[int] = None
     fast_reset: bool = False
     acquisition_type: AcquisitionType = AcquisitionType.DISCRIMINATION
     averaging_mode: AveragingMode = AveragingMode.SINGLESHOT
